refactor: organize assets by type (svg/png), adapt change to the tree
