+++ src/common/components/Icon/index.tsx
@@ -1,12 +1,12 @@
-import { ReactComponent as arrowRight } from "../../../assets/arrow-right.svg";
-import { ReactComponent as cutArrowRight } from "../../../assets/cut-arrow-right.svg";
-import { ReactComponent as ethereum } from "../../../assets/ethereum.svg";
-import { ReactComponent as usdc } from "../../../assets/usdc.svg";
-import { ReactComponent as polygon } from "../../../assets/polygon.svg";
-import { ReactComponent as hop } from "../../../assets/hop.svg";
-import { ReactComponent as copy } from "../../../assets/copy-icon.svg";
-import { ReactComponent as arbitrum } from "../../../assets/arbitrum.svg";
-import { ReactComponent as optimism } from "../../../assets/optimism.svg";
+import { ReactComponent as arrowRight } from "../../../assets/svg/arrow-right.svg";
+import { ReactComponent as cutArrowRight } from "../../../assets/svg/cut-arrow-right.svg";
+import { ReactComponent as ethereum } from "../../../assets/svg/ethereum.svg";
+import { ReactComponent as usdc } from "../../../assets/svg/usdc.svg";
+import { ReactComponent as polygon } from "../../../assets/svg/polygon.svg";
+import { ReactComponent as hop } from "../../../assets/svg/hop.svg";
+import { ReactComponent as copy } from "../../../assets/svg/copy-icon.svg";
+import { ReactComponent as arbitrum } from "../../../assets/svg/arbitrum.svg";
+import { ReactComponent as optimism } from "../../../assets/svg/optimism.svg";
 
 export const icons = {
   arrowRight,
